Split out MARCFetcher
